Add a test for health checks
