+++ proxy.go
@@ -16,152 +16,27 @@ package main
 
 import (
 	"context"
-	"encoding/json"
-	"fmt"
-	"net"
-	"net/http"
 	"os"
 	"os/signal"
-	"strings"
-	"sync"
-	"time"
 
-	"github.com/datastax/cql-proxy/astra"
 	"github.com/datastax/cql-proxy/proxy"
-	"github.com/datastax/cql-proxy/proxycore"
-
-	"github.com/alecthomas/kong"
-	"github.com/datastax/go-cassandra-native-protocol/primitive"
-	"go.uber.org/zap"
 )
 
-const livenessPath = "/liveness"
-const readinessPath = "/readiness"
-
-var cli struct {
-	Bundle             string        `help:"Path to secure connect bundle" short:"b" env:"BUNDLE"`
-	Username           string        `help:"Username to use for authentication" short:"u" env:"USERNAME"`
-	Password           string        `help:"Password to use for authentication" short:"p" env:"PASSWORD"`
-	ContactPoints      []string      `help:"Contact points for cluster. Ignored if using the bundle path option." short:"c" env:"CONTACT_POINTS"`
-	ProtocolVersion    string        `help:"Initial protocol version to use when connecting to the backend cluster (default: v4, options: v3, v4, v5, DSEv1, DSEv2)" short:"n" env:"PROTOCOL_VERSION"`
-	MaxProtocolVersion string        `help:"Max protocol version supported by the backend cluster (default: v4, options: v3, v4, v5, DSEv1, DSEv2)" short:"m" env:"MAX_PROTOCOL_VERSION"`
-	Bind               string        `help:"Address to use to bind server" short:"a" default:":9042" env:"BIND"`
-	Debug              bool          `help:"Show debug logging" env:"DEBUG"`
-	HealthCheck        bool          `help:"Enable liveness and readiness checks" default:"false" env:"HEALTH_CHECK"`
-	HttpBind           string        `help:"Address to use to bind HTTP server used for health checks" default:":8000" env:"HTTP_BIND"`
-	HeartbeatInterval  time.Duration `help:"Interval between performing heartbeats to the cluster" default:"30s" env:"HEARTBEAT_INTERVAL"`
-	IdleTimeout        time.Duration `help:"Duration between successful heartbeats before a connection to the cluster is considered unresponsive and closed" default:"60s" env:"IDLE_TIMEOUT"`
-	ReadinessTimeout   time.Duration `help:"Duration the proxy is unable to connect to the backend cluster before it is considered not ready" default:"30s" env:"READINESS_TIMEOUT"`
-}
-
 func main() {
-	cliCtx := kong.Parse(&cli)
-
-	var resolver proxycore.EndpointResolver
-
-	if len(cli.Bundle) > 0 {
-		bundle, err := astra.LoadBundleZipFromPath(cli.Bundle)
-		if err != nil {
-			cliCtx.Fatalf("unable to open bundle %s: %v", cli.Bundle, err)
-		}
-		resolver = astra.NewResolver(bundle)
-	} else if len(cli.ContactPoints) > 0 {
-		resolver = proxycore.NewResolver(cli.ContactPoints...)
-	} else {
-		cliCtx.Fatalf("must provide either bundle path or contact points")
-	}
-
-	if cli.HeartbeatInterval >= cli.IdleTimeout {
-		cliCtx.Fatalf("idle-timeout must be greater than heartbeat-interval")
-	}
-
-	version := primitive.ProtocolVersion4
-	if len(cli.ProtocolVersion) > 0 {
-		var ok bool
-		if version, ok = parseProtocolVersion(cli.ProtocolVersion); !ok {
-			cliCtx.Fatalf("unsupported protocol version: %s", cli.ProtocolVersion)
-		}
-	}
-
-	maxVersion := primitive.ProtocolVersion4
-	if len(cli.MaxProtocolVersion) > 0 {
-		var ok bool
-		if maxVersion, ok = parseProtocolVersion(cli.MaxProtocolVersion); !ok {
-			cliCtx.Fatalf("unsupported max protocol version: %s", cli.ProtocolVersion)
-		}
-	}
-
-	if version > maxVersion {
-		cliCtx.Fatalf("default protocol version is greater than max protocol version")
-	}
-
 	ctx, cancel := signalContext(context.Background(), os.Interrupt, os.Kill)
 	defer cancel()
 
-	var logger *zap.Logger
-	var err error
-	if cli.Debug {
-		logger, err = zap.NewDevelopment()
-	} else {
-		logger, err = zap.NewProduction()
-	}
-	if err != nil {
-		cliCtx.Fatalf("unable to create logger")
-	}
-
-	var auth proxycore.Authenticator
-
-	if len(cli.Username) > 0 || len(cli.Password) > 0 {
-		auth = proxycore.NewPasswordAuth(cli.Username, cli.Password)
-	}
-
-	p := proxy.NewProxy(ctx, proxy.Config{
-		Version:           version,
-		MaxVersion:        maxVersion,
-		Resolver:          resolver,
-		ReconnectPolicy:   proxycore.NewReconnectPolicy(),
-		NumConns:          1,
-		Auth:              auth,
-		Logger:            logger,
-		HeartBeatInterval: cli.HeartbeatInterval,
-		IdleTimeout:       cli.IdleTimeout,
-	})
-
-	cli.Bind = maybeAddPort(cli.Bind, "9042")
-	cli.HttpBind = maybeAddPort(cli.HttpBind, "8000")
-
-	maybeAddHealthCheck(p)
-
-	cliCtx.FatalIfErrorf(listenAndServe(p, ctx, logger))
-}
-
-func parseProtocolVersion(s string) (version primitive.ProtocolVersion, ok bool) {
-	ok = true
-	lowered := strings.ToLower(s)
-	if lowered == "3" || lowered == "v3" {
-		version = primitive.ProtocolVersion3
-	} else if lowered == "4" || lowered == "v4" {
-		version = primitive.ProtocolVersion4
-	} else if lowered == "5" || lowered == "v5" {
-		version = primitive.ProtocolVersion5
-	} else if lowered == "65" || lowered == "dsev1" {
-		version = primitive.ProtocolVersionDse1
-	} else if lowered == "66" || lowered == "dsev2" {
-		version = primitive.ProtocolVersionDse1
-	} else {
-		ok = false
-	}
-	return version, ok
+	os.Exit(proxy.Run(ctx, os.Args[1:]))
 }
 
 func signalContext(parent context.Context, sig ...os.Signal) (context.Context, func()) {
 	ctx, cancel := context.WithCancel(parent)
-	c := make(chan os.Signal, 1)
-	signal.Notify(c, sig...)
+	ch := make(chan os.Signal)
+	signal.Notify(ch, sig...)
 	if ctx.Err() == nil {
 		go func() {
 			select {
-			case <-c:
+			case <-ch:
 				cancel()
 			case <-ctx.Done():
 			}
@@ -169,117 +44,6 @@ func signalContext(parent context.Context, sig ...os.Signal) (context.Context, f
 	}
 	return ctx, func() {
 		cancel()
-		signal.Stop(c)
+		signal.Stop(ch)
 	}
-}
-
-func maybeAddHealthCheck(p *proxy.Proxy) {
-	if cli.HealthCheck {
-		http.HandleFunc(livenessPath, func(writer http.ResponseWriter, request *http.Request) {
-			writer.WriteHeader(http.StatusOK)
-			_, _ = writer.Write([]byte("ok"))
-		})
-		http.HandleFunc(readinessPath, func(writer http.ResponseWriter, request *http.Request) {
-			header := writer.Header()
-			header.Set("Content-Type", "application/json")
-
-			outageDuration := p.OutageDuration()
-			response, err := json.Marshal(struct {
-				OutageDuration string
-			}{outageDuration.String()})
-			if err != nil {
-				http.Error(writer, fmt.Sprintf("failed to marshal json response: %v", err), http.StatusInternalServerError)
-				return
-			}
-
-			if outageDuration < cli.ReadinessTimeout {
-				writer.WriteHeader(http.StatusOK)
-				_, _ = writer.Write(response)
-			} else {
-				writer.WriteHeader(http.StatusServiceUnavailable)
-				_, _ = writer.Write(response)
-			}
-		})
-	}
-}
-
-//maybeAddPort adds the default port to an IP; otherwise, it returns the original address
-func maybeAddPort(addr string, defaultPort string) string {
-	if net.ParseIP(addr) != nil {
-		return net.JoinHostPort(addr, defaultPort)
-	}
-	return addr
-}
-
-func listenAndServe(p *proxy.Proxy, ctx context.Context, logger *zap.Logger) (err error) {
-	var wg sync.WaitGroup
-
-	ch := make(chan error)
-	server := http.Server{Addr: cli.HttpBind}
-
-	numServers := 1 // Without the HTTP server
-
-	// Listen is called first to set up the listening server connection and establish initial client connections to the
-	// backend cluster so that when the readiness check is hit the proxy is actually ready.
-	err = p.Listen(cli.Bind)
-	if err != nil {
-		return err
-	}
-
-	var listener *net.TCPListener
-
-	if cli.HealthCheck {
-		numServers++ // Add the HTTP server
-
-		tcpAddr, err := net.ResolveTCPAddr("tcp", cli.HttpBind)
-		if err != nil {
-			return err
-		}
-		listener, err = net.ListenTCP("tcp", tcpAddr)
-		if err != nil {
-			return err
-		}
-
-		logger.Info("health checks are listening",
-			zap.String("livenessURL", cli.HttpBind+livenessPath),
-			zap.String("readinessURL", cli.HttpBind+readinessPath))
-	}
-
-	wg.Add(numServers)
-
-	go func() {
-		select {
-		case <-ctx.Done():
-			logger.Debug("proxy interrupted/killed")
-			_ = server.Close()
-			_ = p.Shutdown()
-		}
-	}()
-
-	go func() {
-		defer wg.Done()
-		err = p.Serve()
-		if err != nil {
-			ch <- err
-		}
-	}()
-
-	if cli.HealthCheck {
-		go func() {
-			defer wg.Done()
-			err = server.Serve(listener)
-			if err != nil {
-				ch <- err
-			}
-		}()
-
-	}
-
-	for err = range ch {
-		if err != nil {
-			return nil
-		}
-	}
-
-	return err
 }
